chore: refactor field to use useField hook
